Add Matlab SCons tool.
Added an SCons extension that sets various environment variables by default. The
SConstruct file is now a mere 51 lines vs. 155 for CMake :-) (both with
comments). To do this, however, it starts a Matlab subprocess, which is the only
cross-platform way I could imagine doing this, since it seems Matlab under
Windows doesn't seem to support the "-e" option for getting Matlabs environment
variables.