SConstruct: use variant dirs.
The SCons based build system now uses variant dirs for differentiating
between debug/release builds.  For this a second SConstruct file located in src/
is used.  So far I think this approach is cleaner than using separate
environments in one SConstruct, because the original environment can stay in a
pristine state that is only modified within the secondary SConstruct file.