Add initial Mex file for block-wise reading.
This Mex file works, as long as you are careful.  Initial performance tests show
a reduction of the overhead from opening and closing the file.  This is
especially true for small block sizes: when reading 256 samples, the time to
read the samples was halved.

The following caveats apply for now:
- error handling needs to be improved
- only a single file can open for block-wise reading at a time