Reorganise performance test code.
The performance test code has been completely reorganized and cleaned up.  The
individual tests now reside in four functions in the test.perf namespace, each
of which returns the mean and STD of the measured run times and a cell array of
labels.  These functions are called by the script test.performance, which also
generates and prints the plots.