msndfile.read: auto-add extension for most files.
The file type extension is now automatically added for what libsndfile refers to
as simple formats (as documented in the README, see diff).  If there is an
ambiguity, msndfile.read returns an error, unless there is a WAV file, in which
case it reads that.  For all other cases, a file type extension must be
provided.