Return empty matrix when 0 frames are read.
Matlab 2010a returns the entire file when N=0, however, it prints a warning that
in the future, an empty matrix will be returned instead. Msndfile does neither,
and the latter was simpler to implement (and makes more sense), so
msndfile.read() and .blockread() will now return an empty 0xnum_channels matrix
whenever 0 frames are read (even if it is unintentional).