Use mxArrayToString() to get the file name.
Migrate from mxGetString() to mxArrayToString() for getting the file name, which
is needed for multibyte file name support.  This means that up to now msndfile
could not open files with non-ASCII characters in the name, of which I was
heretofore unaware.  Oh well, it should work now (it does for me).