Fix multi-byte file name tests on non-UTF8 systems.
Since Matlab reads m-Files using the default system encoding, the multi-byte
file name tests failed when run on a non-UTF8 system, since they were stored as
UTF-8 in the m-File but are interpreted using a different character set.

This is now worked around by storing part of the file name as a UTF-8 encoded
array of bytes (a uint8 array) and then converting that to unicode via
native2unicode().