SConstruct: update to work correctly on Windows.
Update the SCons based build system to work correctly on Windows.  This entails
overriding the default tools on Windows to not include the Microsoft tools,
because msndfile requires C99 support and the Visual Compiler does not (and most
likely will not ever) support that.

Furthermore, most of the Windows/Unix special-casing has become unnecessary,
including the "--force-mingw" option, which has therefor been removed.