Fix two cases of initialising a zero-length array
These two errors were caught by scan-build, which is distributed with clang.
In both cases, conditions may arise in which a VLA is initialised with length
zero.  Neither array is actually *used* afterwards (other checks will lead to
an exception first), and have not — on my system at least — lead to any
compiler diagnostics or crashes, but still: strictly speaking, an array must
have at least one element.

Note that there are still a number of reports by scan-build, for blockread.c
and read_utils.c.  In the case of the former, all reports are clearly false
positives, because scan-build does not know that the mexErr* functions exit the
function.  In the case of the latter, I'm not 100% sure, but both reports seem
to assume that file_exts only has one element, which is not true.  I think that
scan-build cannot take into account that num_formats depends on an external
function, or that num_formats is the same in get_format_extensions() and
gen_filename(), or both.